rename drug types
Entire-Checkpoint: 68282d86cae2
