+++ tbsim/interventions/__init__.py
@@ -9,7 +9,7 @@ from .interventions import *
 from .beta import *
 from .tpt import *
 from .bcg import *
-from .tb_drug_types import *
+from .drug_types import *
 from .tb_health_seeking import *
 from .dx_products import *
 from .tx_products import *
